requirements.txt
tensorflow
keras
matplotlib
numpy
scikit-learn
seaborn
Pillow